requirements.txt
# Core Data Analysis
pandas>=2.0.0
numpy>=1.24.0

# Database Connectors
mysql-connector-python>=8.0.0
psycopg2-binary>=2.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Dashboard
streamlit>=1.25.0

# REST API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Machine Learning & Analytics
scikit-learn>=1.3.0
scipy>=1.11.0

# Configuration & Security
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0

# Data Export
openpyxl>=3.1.0
